refactor: fail-fast load, extract loader & pagination, immutable ZipCode
Ronda de mejoras de robustez, limpieza y estructura. La suite completa
(66 tests) permanece verde.

Robustez / config:
- ZipCodeService: la carga es fail-fast — si el catálogo no se encuentra
  o queda vacío se aborta el arranque (IllegalStateException) en vez de
  dejar el servicio vivo-pero-roto (NPE en /federal-entities, 200 con
  cuerpo nulo en /stats).
- prod/railway: deshabilita springdoc api-docs y swagger-ui.
- qa: elimina spring.cache.caffeine.spec (ignorado por registerCustomCache).
- Añade banner-dev.txt (estaba referenciado pero no existía).

Limpieza:
- Elimina los métodos @Deprecated de 1 argumento (search*/advancedSearch),
  el helper findOrderedCandidatesInIndex y las 3 regiones de caché no
  paginadas; tests migrados a las variantes paginadas.
- Quita el @ExceptionHandler(IOException) inalcanzable.
- Quita el @Cacheable redundante de getAllFederalEntities + su región.

Estructura:
- ZipCode es inmutable (@Builder, campos final, sin @Setter).
- Extrae carga/parseo a SepomexZipCodeLoader -> ZipCodeData y la
  paginación a PaginationSupport; ZipCodeService queda enfocado en
  consultas (935 -> ~430 líneas) y publica los índices como volatile.

Docs:
- CLAUDE.md sincronizado (Spring Boot 4.0.6, 66 tests, HashMap/TreeMap,
  regiones de caché reales, ZipCodeApi/CacheControlInterceptor).
