+++ src/main/java/com/coderalexis/CodigoPostalApi/exceptions/GlobalExceptionHandler.java
@@ -11,7 +11,6 @@ import org.springframework.web.servlet.resource.NoResourceFoundException;
 import jakarta.validation.ConstraintViolationException;
 import lombok.extern.slf4j.Slf4j;
 
-import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.stream.Collectors;
 
@@ -110,20 +109,5 @@ public class GlobalExceptionHandler {
         );
         log.warn("IllegalArgumentException: {}", ex.getMessage());
         return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(IOException.class)
-    public ResponseEntity<ErrorResponse> handleIOException(
-            IOException ex,
-            WebRequest request
-    ) {
-        ErrorResponse error = new ErrorResponse(
-                HttpStatus.SERVICE_UNAVAILABLE.value(),
-                "Error al acceder a los datos",
-                "El servicio no puede acceder a los datos en este momento",
-                LocalDateTime.now()
-        );
-        log.error("IOException: ", ex);
-        return new ResponseEntity<>(error, HttpStatus.SERVICE_UNAVAILABLE);
     }
 }
