+++ src/main/java/com/coderalexis/CodigoPostalApi/model/ZipCode.java
@@ -2,40 +2,45 @@ package com.coderalexis.CodigoPostalApi.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
 import lombok.ToString;
 
 import java.util.List;
 
 /**
- * NOTA sobre @JsonInclude: la inclusión NON_NULL ya está configurada
+ * Código postal con sus asentamientos. Es <strong>inmutable</strong> (#11): se
+ * construye una sola vez durante la carga del catálogo (vía {@code @Builder}) y
+ * sus campos no cambian después. Esto elimina cualquier posibilidad de mutación
+ * accidental del estado interno una vez publicado en los índices.
+ *
+ * <p>NOTA sobre @JsonInclude: la inclusión NON_NULL ya está configurada
  * globalmente en application.yml (spring.jackson.default-property-inclusion),
- * así que no hace falta repetirla aquí.
+ * así que no hace falta repetirla aquí.</p>
  */
 @Getter
-@Setter
 @ToString
-@NoArgsConstructor
+@Builder
+@AllArgsConstructor
 @EqualsAndHashCode(of = "zipCode")
 public class ZipCode {
     @JsonProperty("zip_code")
-    private String zipCode;
+    private final String zipCode;
 
-    private String locality;
+    private final String locality;
 
     @JsonProperty("federal_entity")
-    private String federalEntity;
+    private final String federalEntity;
 
-    private String municipality;
+    private final String municipality;
 
-    private List<Settlements> settlements;
-
-    @JsonIgnore
-    private String normalizedFederalEntity;
+    private final List<Settlements> settlements;
 
     @JsonIgnore
-    private String normalizedMunicipality;
+    private final String normalizedFederalEntity;
+
+    @JsonIgnore
+    private final String normalizedMunicipality;
 }
